requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
boto3==1.24.46
botocore==1.27.46
Django==4.0.6
django-environ==0.9.0
django-storages==1.13
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
jmespath==1.0.1
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.1
s3transfer==0.6.0
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.2.0
toml==0.10.2
urllib3==1.26.11
